+++ app/DefaultApp/Vues/block/liste_imagerie.php
@@ -1,7 +1,15 @@
 <div class="card-body">
     <h4>Examens imagerie disponible</h4>
-    <a href="imagerie?administration&examens&ajouter" class="btn btn-primary btn-xs">Ajouter</a>
-    <a href="imagerie?administration&examens" class="btn btn-warning btn-xs">Lister</a>
+    <?php
+    $role=\systeme\Model\Utilisateur::role();
+    if($role=="admin"){
+        ?>
+        <a href="imagerie?administration&examens&ajouter" class="btn btn-primary btn-xs">Ajouter</a>
+        <a href="imagerie?administration&examens" class="btn btn-warning btn-xs">Lister</a>
+    <?php
+    }
+    ?>
+
     <hr>
     <?php
     if (isset($_GET['ajouter'])) {
@@ -46,10 +54,16 @@
                             <td><?= $img->getCout() ?></td>
                             <td><?= $img->getPrix() ?></td>
                             <th>
-                                <a class="btn btn-primary btn-xs"
-                                   href="imagerie?administration&examens&modifier=<?= $img->id ?>">Modifier</a>
-                                <a class="btn btn-danger btn-xs"
-                                   href="imagerie?administration&examens&supprimer=<?= $img->id ?>">Supprimer</a>
+                                <?php
+                                if($role=="admin"){
+                                    ?>
+                                    <a class="btn btn-primary btn-xs"
+                                       href="imagerie?administration&examens&modifier=<?= $img->id ?>">Modifier</a>
+                                    <a class="btn btn-danger btn-xs"
+                                       href="imagerie?administration&examens&supprimer=<?= $img->id ?>">Supprimer</a>
+                                    <?php
+                                }
+                                ?>
                             </th>
                         </tr>
                         <?php
